% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addInfo}
\alias{addInfo}
\title{adding additional info to data frame from info_df}
\usage{
addInfo(data_df = NULL, info_df = NULL)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
addInfo
}
\note{
todo -
}
\keyword{plate}
\keyword{readers}

