% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalVis.R
\name{as.numeric.factor}
\alias{as.numeric.factor}
\title{as.numeric.factor}
\usage{
\method{as.numeric}{factor}(x)
}
\value{
numeric vector
}
\description{
converting factor to factor
}
\details{
as.numeric.factor
}
