% Generated by roxygen2 (4.0.2): do not edit by hand
\name{barplot3d}
\alias{barplot3d}
\title{_____________________________________________________________________________
barplot3d}
\usage{
barplot3d(heights_arr = NULL, barcode = "0000", filename = "_3Dbarplot",
  column_colours = "bluered", numCol = 20, zmax = 1, nticks = 4,
  moreLight = FALSE, axisBack = TRUE, theta = -60, phi = 15,
  scale = 1, char_scale = 1.1, width = 700, height = 900, transp = 1,
  col_lab = NULL, row_lab = NULL, z_lab = NULL, barWidth = 0.5,
  barDistance = 0.8, plotBg = "white", outputFormat = "screen")
}
\arguments{
\item{heigths_arr}{}
}
\description{
plots data from a (heights x rows x columns ) array
}
\details{
Plotting 3D Bars
}
\note{
todo:
}
\examples{
barplot3d()
}
\keyword{plate}
\keyword{readers}

