% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calcAllLinModels}
\alias{calcAllLinModels}
\title{Calculate All Linear Models of given Data Frame}
\usage{
calcAllLinModels(kin_data_df = NULL, barcode = "0000", wavelength = 0,
  modelVar = c("DiffTime", "Value"), groupingVar = "Well")
}
\arguments{
\item{kin_data_df}{(data.frame)}

\item{wavelength}{= 0 (int)}

\item{barcode="0000"}{(string)}
}
\value{
data frame with slopes added
}
\description{
Calculates all possible linar models and stores coefficients in a data frame with the following columns
             'Well', 'Barcode', 'Num', 'Type',  'Description', 'Slope', 'Intercept'
}
\details{
calcAllLinModels
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: adding R-square to output
}
\examples{
lin_model_df <- calcAllLinModels(kin_data_df, wavelength=245)
}
\keyword{plate}
\keyword{readers}

