% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linModels.R
\name{datasetMean}
\alias{datasetMean}
\title{Mean of each individual well in a dataset}
\usage{
datasetMean(kin_data_df = NULL, barcode = "0000", wavelength = 0,
  modelVar = c("DiffTime", "Value"), groupingVar = "Well",
  minIntervalSize = 0, minIntervalSizeFraction = 0.8)
}
\arguments{
\item{modelVar=c("DiffTime", }{"Value")}
}
\value{
data frame with
}
\description{
Calculates the mean for each well in a container
}
\details{
datasetMean
}
\note{
todo: Rewrite as aggregate function ?
      system.time()
}
\examples{
 for examples, please wait until fixed !!!
}
\keyword{plate}
\keyword{readers}

