% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_Plot.chromatogram}
\alias{LA_Plot.chromatogram}
\title{chromatogram Plot}
\usage{
\method{LA_Plot}{chromatogram}(plot_data_df = NULL, plotVar = c("DiffTime",
  "Value"), groupingVar = "Type", plot_type = "l", single_plot = TRUE,
  plot_overview = FALSE, curr_filename = "_chromatogram_plot", ...)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
LA_Plot.chromatogram
}
\note{
todo - time conversion (h, min)
}
\keyword{plate}
\keyword{readers}

