% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laraPlot.R
\name{LA_Plot.core}
\alias{LA_Plot.core}
\title{core plate plot function}
\usage{
\method{LA_Plot}{core}(plot_data_df = NULL, slopes_df = NULL,
  fl_df = NULL, fit_lst = NULL, norm_df = NULL, peaksIdx = NULL,
  num_col = 10, ...)
}
\arguments{
\item{plot_data_df}{}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
LA_Plot.core
}
\note{
todo - colour palette, scatterplot, single/overview fix, all_meta_data_df (slopes, fl)
}
\keyword{plate}
\keyword{readers}
