% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_Plot.default}
\alias{LA_Plot.default}
\title{Plate Plot functions - default function}
\usage{
\method{LA_Plot}{default}(plot_data_df = NULL, ...)
}
\arguments{
\item{plot_data_df}{}
}
\description{
Default function
}
\details{
LA_Plot.default
}

