% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinPlot.R
\name{LA_Plot.growthRibbon}
\alias{LA_Plot.growthRibbon}
\title{Plotting bacterial growth and Expression Data}
\usage{
\method{LA_Plot}{growthRibbon}(plot_data_df = NULL, wavelength = 600,
  numGrowth = 0, plot_type = "b", timeUnit = "mins", sampleType = "S",
  averageGrowth = FALSE, plotVar = c("DiffTimeUnit", "Value"),
  groupingVar = "Well", plot_description = TRUE, plot_well_names = TRUE,
  plotTitle = "Av. Growth and Expression of Plate ",
  filename = "_growthExprPlotRibbon", ...)
}
\description{
Slopes need to be pre-calculated to also plot max and min slopes
}
\details{
LA_Plot.growthRibbon
}
\note{
todo: not working yet - just a concept
}
\examples{
   good parameters f. jpg: plotWidth=7600, plotHeight=4400, plotQuality=70
   single plot: plotWidth=1200, plotHeight=800, plotQuality=70
}
\author{
Johannes Kabisch (big thanx Johannes !)
}
\keyword{plate}
\keyword{readers}

