% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{LA_Plot.histogram}
\alias{LA_Plot.histogram}
\title{histogram plot}
\usage{
\method{LA_Plot}{histogram}(plot_data_df = NULL, breaks = 10, ...)
}
\arguments{
\item{breaks}{- number of breaks for histogram to}
}
\description{
Plots a histogram of plotVar with n breaks
}
\details{
LA_Plot.histogram
}
\note{
todo
}
\author{
Mark Doerr (mark.doerr at uni-greifswald.de)
}
\keyword{plate}
\keyword{readers}
