% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinPlot.R
\name{LA_Plot.kinetics}
\alias{LA_Plot.kinetics}
\title{kinetic Plot}
\usage{
\method{LA_Plot}{kinetics}(plot_data_df = NULL, slopes_df = NULL,
  fl_df = NULL, plotVar = c("DiffTime", "Value"), timeUnit = "secs",
  groupingVar = "Well", plot_well_names = TRUE, description = TRUE,
  single_plot = FALSE, plot_overlay = FALSE, plot_overview = TRUE,
  filename = "_kinPlot", ...)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction 
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
LA_Plot.kinetics
}
\note{
todo - time conversion (h, min)
}
\keyword{plate}
\keyword{readers}
