% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_Plot.MMkinetics}
\alias{LA_Plot.MMkinetics}
\title{kinetic Plot}
\usage{
\method{LA_Plot}{MMkinetics}(plot_data_df = NULL, fit_lst = NULL,
  single_plot = TRUE, overlay = TRUE, plot_fitting = TRUE,
  plotType = "p", plot_xlab = "Substr. Conc / M",
  plot_ylab = "velocity / M s-1", groupingVar = "Enzyme",
  plotVar = c("Substrate1Conc", "v0"), filename = "MM_kinPlot", ...)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
LA_Plot.MMkinetics
}
\note{
todo - time conversion (h, min)
}
\keyword{plate}
\keyword{readers}

