% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinPlot.R
\name{LA_Plot.pHProfile}
\alias{LA_Plot.pHProfile}
\title{kinetic Plot}
\usage{
\method{LA_Plot}{pHProfile}(plot_data_df = NULL, fit_lst = NULL,
  single_plot = TRUE, overlay = TRUE, plot_fitting = TRUE,
  plot_type = "p", plot_xlab = "pH", plot_ylab = "Kcat/Km  [1/M 1/s]",
  plotVar = c("MediumpH", "KcatKm"), groupingVar = "Enzyme",
  filename = "_pHProfile_1", ...)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction 
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
LA_Plot.pHProfile
}
\note{
todo - time conversion (h, min)
}
\keyword{plate}
\keyword{readers}

