% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_Plot}
\alias{LA_Plot}
\title{Plate Plot functions - generic function}
\usage{
LA_Plot(plot_data_df = NULL, plotVar = NULL, groupingVar = NULL,
  barcode = "0000", wavelength = 0, overview = FALSE, singlePlot = TRUE,
  overlay = FALSE, preview = FALSE, errorBars = FALSE, fitting = FALSE,
  plotPeaks = FALSE, wellNames = FALSE, description = FALSE,
  markSegments = TRUE, markAverage = TRUE, markBest = FALSE,
  bestNegSlope = FALSE, thresholds = c(1, 1, 1), plotRef = FALSE,
  plotBestLinMod = FALSE, plotType = "b", plotChar = 0, charScale = 0,
  cex = 0, lineType = 1, lineWidth = 1, lineCol = "red",
  fillCol = "lightgrey", numCol = 20, customCol = FALSE,
  panelGeometry = c(8, 12), plotTitle = "", plotLegend = FALSE,
  ylim = NULL, xlim = NULL, plotAxes = TRUE, xlab = "", ylab = "",
  zlab = "", plotGrid = FALSE, plotBg = "white", transp = 1,
  barWidth = 0.5, barDistance = 0.8, theta = -60, phi = 15, scale = 1,
  plotWidth = 1200, plotHeight = 1024, plotQuality = 70,
  filename = "LA_Plot", outputFormat = "png", ...)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
_____________________________________________________________________________
LA_Plot
}
\note{
todo - removing singlePlot concept - allowing only overview or singlePlot
}
\keyword{plate}
\keyword{readers}

