% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralPlot.R
\name{LA_Plot.spectrum}
\alias{LA_Plot.spectrum}
\title{kinetic Plot}
\usage{
\method{LA_Plot}{spectrum}(plot_data_df = NULL, plotVar = c("Wavelength",
  "Value"), groupingVar = "Well", plot_type = "l",
  filename = "_spectralPlot", ...)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction 
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
LA_Plot.spectrum
}
\note{
todo - time conversion (h, min)
}
\keyword{plate}
\keyword{readers}

