% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_Plot.wells}
\alias{LA_Plot.wells}
\title{kinetic Plot}
\usage{
\method{LA_Plot}{wells}(plot_data_df = NULL, colourScheme = "custom",
  plotAxis = TRUE, plotGrid = FALSE, plotBox = FALSE, xlim = c(0, 15),
  ylim = c(0, 3.1), ...)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
LA_Plot.wells
}
\note{
todo -
}
\keyword{plate}
\keyword{readers}

