% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzymeKinetics.R
\name{michaelisMentenKineticsSP}
\alias{michaelisMentenKineticsSP}
\title{converts time units}
\usage{
michaelisMentenKineticsSP(mm_df = NULL)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction 
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
michaelisMentenKineticsSP
}
\note{
todo -
}
\keyword{plate}
\keyword{readers}
