% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enzymeAssays.R
\name{phenolRedAssay}
\alias{phenolRedAssay}
\title{converts time units}
\usage{
phenolRedAssay(assay_df = NULL, intensityCoding = FALSE, limOrange = 0.95,
  limYellow = 0.8, output = FALSE)
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}

\item{use_db_layout=FALSE}{(boolean) - reading the layout from a database}

\item{PLC=FALSE}{(boolean) - automatically calculate pathlength correction 
under the assumptions of a 96 round well SBS layout (r=034 cm)
and 200ul liquid volume}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
phenolRedAssay
}
\note{
todo -
}
\keyword{plate}
\keyword{readers}

