% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readerDF2Array}
\alias{readerDF2Array}
\title{Converts a data frame to an easy plottable 3D array}
\usage{
readerDF2Array(source_data_df = NULL, num = 0, nrows = 8, ncols = 12,
  wavelength = 0, groupingVar = "Well", plotVar = c("Value", "RefTime"),
  add_zero = TRUE)
}
\arguments{
\item{reader_data_df}{(data.frame) - source data frame}

\item{num=0}{(integer) - number of measurements}

\item{wavelength=0}{}

\item{add_zero=TRUE}{}

\item{use_slope=FALSE}{}
}
\value{
array (12 x 8 x num)
}
\description{
This convenient function can be used to transform a plate reader dataframe into an 3D array
             in the format 12 x 8 x num.
             It is used to quickly iterate through kinetic data for each well.
}
\details{
readerDF2Array
}
\note{
todo - arbitrary number of cols and rows
}
\keyword{plate}
\keyword{readers}

