% Generated by roxygen2 (4.0.2): do not edit by hand
\name{selectBestLinModelList}
\alias{selectBestLinModelList}
\title{Select Best Linear Model for a given Data Frame}
\usage{
selectBestLinModelList(kin_data_lst = NULL, barcode = "0000",
  wavelength = 0, minIntervalSize = 0, minIntervalSizeFraction = 0.8)
}
\arguments{
\item{kin_data_df}{}
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
             based on a suggestion of Martin Weiss to find the smallest R-square
}
\details{
selectBestLinModelList
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: not yet adapted to new data frame structure !!
      system.time()
}
\examples{
please wait until fixed !!!
}
\keyword{plate}
\keyword{readers}

