% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trim.leading}
\alias{trim.leading}
\title{trim.leading}
\usage{
trim.leading(x)
}
\value{
returns string w/o leading whitespace
}
\description{
removing leading whitespace from a string
}
\details{
trim.leading
}
\author{
f3lix at stackoverflow (big thanx !)
}

