% Generated by roxygen2 (4.0.2): do not edit by hand
\name{convertEnzymeUnit}
\alias{convertEnzymeUnit}
\title{converts enzyme units}
\usage{
convertEnzymeUnit(unit = "U")
}
\arguments{
\item{unit}{}
}
\description{
This function is the main dispatcher function for reading several file formats into dataframes.
}
\details{
convertEnzymeUnit
}
\note{
todo - not working yet, just a concept
}
\author{
Anders Knight, Mark Doerr
}
\keyword{plate}
\keyword{readers}

