% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitConvert.R
\name{convertTimeUnit}
\alias{convertTimeUnit}
\title{converts time units}
\usage{
convertTimeUnit(unit = "min")
}
\arguments{
\item{layout_dir="./"}{(string) - plate layout file directory with leading /}
}
\description{
This function converts the time
}
\details{
convertTimeUnit
}
\note{
todo -
}
\keyword{plate}
\keyword{readers}
