% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_Plot.barplot}
\alias{LA_Plot.barplot}
\title{barplot plot}
\usage{
\method{LA_Plot}{barplot}(plot_data_df = NULL, topN = 0, ...)
}
\arguments{
\item{topN}{- integer to plot the n best values}
}
\description{
universal bar plot function with plotting mean and standard devaiation, if topN > 0 is provided, it orders the values in descending
             order and plots the first N values
}
\details{
LA_Plot.barplot
}
\note{
todo
}
\author{
Mark Doerr (mark.doerr at uni-greifswald.de)
}
\keyword{plate}
\keyword{readers}

