% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LA_Plot.barplot3d}
\alias{LA_Plot.barplot3d}
\title{Plotting 3D bars of Measurements in Microtiter Format}
\usage{
\method{LA_Plot}{barplot3d}(plot_data_df = NULL, num = 0, zmax = 1, ...)
}
\arguments{
\item{heigths_arr}{}
}
\description{
Wrapper function for 3D bar plotting
}
\details{
LA_Plot.barplot3d
}
\note{
todo:
}
\examples{
platePlot3d()
}
\keyword{plate}
\keyword{readers}

