% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laraPlot.R
\name{LA_Plot.default}
\alias{LA_Plot.default}
\title{Plate Plot functions - default function}
\usage{
\method{LA_Plot}{default}(plot_data_df = NULL, ...)
}
\description{
Default function
}
\details{
LA_Plot.default
}
