% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinPlot.R
\name{LA_Plot.growth}
\alias{LA_Plot.growth}
\title{Plotting bacterial growth and Expression Data}
\usage{
\method{LA_Plot}{growth}(plot_data_df = NULL, wavelength = 660,
  numGrowth = 0, plot_type = "b", timeUnit = "hours", sampleType = "S",
  averageGrowth = FALSE, plotVar = c("DiffTimeUnit", "Value"),
  groupingVar = "Well", plot_description = TRUE, plot_well_names = TRUE,
  plotTitle = "Av. Growth and Expression of Plate ",
  filename = "_growthExprPlot", ...)
}
\description{
Slopes need to be pre-calculated to also plot max and min slopes
}
\details{
LA_Plot.growth
}
\note{
todo: connection line growth expression
}
\examples{
   good parameters f. jpg: plotWidth=7600, plotHeight=4400, plotQuality=70
   single plot: plotWidth=1200, plotHeight=800, plotQuality=70
}
\keyword{plate}
\keyword{readers}

