% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laraEvalVis-package.r
\docType{package}
\name{laraEvalVis}
\alias{laraEvalVis}
\alias{laraEvalVis-package}
\title{A library for data evaluation and visualisation (2D and 3D)}
\details{
A library for data evaluation and visualisation (2D and 3D)
}
\note{
date: 150601

last modification:
}
\author{
mark doerr \email{mark.doerr@uni-greifswald.de},
  and \email{mark@ismeralda.org}
}
