% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linModels.R
\name{selectBestLinModel}
\alias{selectBestLinModel}
\title{Select Best Linear Model for a given Data Frame}
\usage{
selectBestLinModel(kin_data_df = NULL, barcode = "0000", wavelength = 0,
  modelVar = c("DiffTime", "Value"), groupingVar = "Well",
  minIntervalSize = 0, minIntervalSizeFraction = 0.8)
}
\value{
data frame with
}
\description{
Calculates all possible linar models and stores coefficients in a data frame
             based on a suggestion of Martin Weiss to find the smallest R-square
}
\details{
selectBestLinModel
}
\note{
relativly well runtime optimized version - don't be confused about the ugly code: it's quite fast

todo: Rewrite as aggregate function ?
      system.time()
}
\examples{
 for examples, please wait until fixed !!!
}
\keyword{plate}
\keyword{readers}
