% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalVis.R
\name{trim.both}
\alias{trim.both}
\title{trim.both}
\usage{
trim.both(x)
}
\value{
returns string w/o leading or trailing whitespace
}
\description{
removing whitespace from start and end of a string
}
\details{
trim.both
}
\author{
f3lix at stackoverflow (big thanx !)
}

