% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalVis.R
\name{trim.trailing}
\alias{trim.trailing}
\title{trim.trailing}
\usage{
trim.trailing(x)
}
\value{
returns string w/o trailing whitespace
}
\description{
removing trailing whitespace from a string
}
\details{
trim.trailing
}
\author{
f3lix at stackoverflow (big thanx !)
}

