% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTSanalysis.R
\name{ZfactorAnalysis}
\alias{ZfactorAnalysis}
\title{Z-factorAnalysis}
\usage{
ZfactorAnalysis(assay_df = NULL, sampleTypes = c("S", "nCTRL"))
}
\arguments{
\item{a}{}
}
\description{
This function determines the Z-factor.
}
\details{
ZfactorAnalysis
}
\note{
todo -
}
\keyword{plate}
\keyword{readers}
